requirements.txt
botbuilder-integration-aiohttp>=4.15.0

redis
